+++ examples/context_sharing.rs
@@ -73,7 +73,7 @@ impl InputContext for PlayerBox {
 
         ctx.bind::<Move>()
             .with_wasd()
-            .with_modifier(Normalize)
+            .with_modifier(DeadZone::default())
             .with_modifier(DeltaLerp::default())
             .with_modifier(Scale::splat(DEFAULT_SPEED));