+++ examples/local_multiplayer.rs
@@ -110,7 +110,7 @@ impl InputContext for PlayerBox {
         // Can be called multiple times extend bindings.
         // In our case we cant to add modifiers for all players.
         ctx.bind::<Move>()
-            .with_modifier(Normalize)
+            .with_modifier(DeadZone::default())
             .with_modifier(DeltaLerp::default())
             .with_modifier(Scale::splat(DEFAULT_SPEED));